Just remove stdcall support and simplify all the hooking code from it

+++ src/LuaInject/DebugBackend.h
@@ -76,20 +76,20 @@ public:
     /**
      * Attaches the debugger to the state.
      */
-    VirtualMachine* AttachState(unsigned long api, lua_State* L);
+    VirtualMachine* AttachState(int api, lua_State* L);
     
-    void VMInitialize(unsigned long api, lua_State* L, VirtualMachine* vm);
+    void VMInitialize(int api, lua_State* L, VirtualMachine* vm);
 
     /**
      * Detaches the debugger from the state.
      */
-    void DetachState(unsigned long api, lua_State* L);
+    void DetachState(int api, lua_State* L);
 
     /**
      * Sends information about the script to the front end. This is called after
      * calling the lua_load (or similar) method.
      */
-    int PostLoadScript(unsigned long api, int result, lua_State* L, const char* source, size_t size, const char* name);
+    int PostLoadScript(int api, int result, lua_State* L, const char* source, size_t size, const char* name);
 
     /**
      * Registers a script with the backend. This will tell track this source file
@@ -100,7 +100,7 @@ public:
      */
     int RegisterScript(lua_State* L, const char* source, size_t size, const char* name, bool unavailable);
 
-    int RegisterScript(unsigned long api, lua_State* L, lua_Debug* ar);
+    int RegisterScript(int api, lua_State* L, lua_Debug* ar);
 
     /**
      * Steps execution of a "broken" script by one line. If the current line
@@ -131,7 +131,7 @@ public:
      * Evalates the expression. If there was an error evaluating the expression the
      * method returns false and the error message is stored in the result.
      */
-    bool Evaluate(unsigned long api, lua_State* L, const std::string& expression, int stackLevel, std::string& result);
+    bool Evaluate(int api, lua_State* L, const std::string& expression, int stackLevel, std::string& result);
 
     /**
      * Evalates the expression. If there was an error evaluating the expression the
@@ -148,7 +148,7 @@ public:
     /**
      * Gets the value of a variable. The variable can include member selection.
      */
-    TiXmlNode* GetValue(unsigned long api, lua_State* L, const std::string& variable, unsigned int scriptIndex,
+    TiXmlNode* GetValue(int api, lua_State* L, const std::string& variable, unsigned int scriptIndex,
         unsigned int line);
 
     /**
@@ -170,7 +170,7 @@ public:
      * Calls the function on the top of the stack in a protected environment that
      * triggers a debugger exception on error.
      */
-    int Call(unsigned long api, lua_State* L, int nargs, int nresults, int errorfunc);
+    int Call(int api, lua_State* L, int nargs, int nresults, int errorfunc);
 
     /**
      * Returns the index of the script in the scripts array with the specified
@@ -179,25 +179,25 @@ public:
      */
     int GetScriptIndex(const char* name) const;
 
-    bool StackHasBreakpoint(unsigned long api, lua_State* L);
+    bool StackHasBreakpoint(int api, lua_State* L);
 
     /**
      * Returns the class name associated with the metatable index. This makes
      * a few assumptions, namely that the metatable was associated with a global
      * variable which is the class name (i.e. what luaL_newmetatable does).
      */
-    bool GetClassNameForMetatable(unsigned long api, lua_State* L, int mt) const;
+    bool GetClassNameForMetatable(int api, lua_State* L, int mt) const;
 
     /**
      * Pushes onto the stack the class name for a userdata.
      */
-    const char* GetClassNameForUserdata(unsigned long api, lua_State* L, int ud) const;
+    const char* GetClassNameForUserdata(int api, lua_State* L, int ud) const;
 
     /**
      * Called to register a metatable with a class name. This allows the lookup
      * of class names based on the userdata object's metatable.
      */
-    void RegisterClassName(unsigned long api, lua_State* L, const char* name, int metaTable);
+    void RegisterClassName(int api, lua_State* L, const char* name, int metaTable);
 
     /**
      * Sends a text message to the front end.
@@ -218,19 +218,19 @@ public:
      * Callback from Lua when a debug event (new line, function enter or exit)
      * occurs.
      */
-    void HookCallback(unsigned long api, lua_State* L, lua_Debug* ar);
+    void HookCallback(int api, lua_State* L, lua_Debug* ar);
 
     /**
      * Called when a new API is created.
      */
-    void CreateApi(unsigned long api);
+    void CreateApi(int api);
 
     /**
      * Enables or disable just-in-time compilation for the state. If LuaJIT is not being used
      * this has no effect and returns false. If JIT was successfully enabled or disabled the 
      * function returns true.
      */
-    bool EnableJit(unsigned long api, lua_State* L, bool enable);
+    bool EnableJit(int api, lua_State* L, bool enable);
 
 private:
 
@@ -297,12 +297,12 @@ private:
      * Breaks from inside the script code. This will block until execution
      * is resumed.
      */
-    void BreakFromScript(unsigned long api, lua_State* L);
+    void BreakFromScript(int api, lua_State* L);
 
     /**
      * Error handling call back for relaying exceptions.
      */
-    int ErrorHandler(unsigned long api, lua_State* L);
+    int ErrorHandler(int api, lua_State* L);
 
     /**
      * Static version of the error handler that just forwards to the non-static
@@ -316,7 +316,7 @@ private:
      * be skipped. This is usful when the current execution point is an error
      * handler we defined.
      */
-    void SendBreakEvent(unsigned long api, lua_State* L, int stackTop = 0);
+    void SendBreakEvent(int api, lua_State* L, int stackTop = 0);
 
     /**
      * Sends an exception event to the frontend. Break events should be sent
@@ -334,19 +334,19 @@ private:
      * Gets the value at location n on the stack as text. If expandTable is true
      * then tables will be returned in their expanded form (i.e. "{ ... }")
      */
-    TiXmlNode* GetValueAsText(unsigned long api, lua_State* L, int n, int maxDepth = 10, const char* typeNameOverride = NULL, bool displayAsKey = false) const;
+    TiXmlNode* GetValueAsText(int api, lua_State* L, int n, int maxDepth = 10, const char* typeNameOverride = NULL, bool displayAsKey = false) const;
 
     /**
      * Gets the value at location n on the stack as text. If expandTable is true
      * then tables will be returned in their expanded form (i.e. "{ ... }")
      */
-    TiXmlNode* GetLuaBindClassValue(unsigned long api, lua_State* L, unsigned int maxDepth, bool displayAsKey = false) const;
+    TiXmlNode* GetLuaBindClassValue(int api, lua_State* L, unsigned int maxDepth, bool displayAsKey = false) const;
 
     /**
      * Gets the table value at location n on the stack as text. Nested tables are
      * not expanded.
      */
-    TiXmlNode* GetTableAsText(unsigned long api, lua_State* L, int t, int maxDepth = 10, const char* typeNameOverride = NULL) const;
+    TiXmlNode* GetTableAsText(int api, lua_State* L, int t, int maxDepth = 10, const char* typeNameOverride = NULL) const;
 
     /**
      * Returns true if the name belongs to a Lua internal variable that we
@@ -357,12 +357,12 @@ private:
     /**
      * This is semantically equivalent to luaL_loadstring.
      */
-    int LoadScriptWithoutIntercept(unsigned long api, lua_State* L, const char* string, size_t size, const char* name);
+    int LoadScriptWithoutIntercept(int api, lua_State* L, const char* string, size_t size, const char* name);
 
     /**
      * This is semantically equivalent to luaL_loadstring.
      */
-    int LoadScriptWithoutIntercept(unsigned long api, lua_State* L, const std::string& string);
+    int LoadScriptWithoutIntercept(int api, lua_State* L, const std::string& string);
     
     /**
      * Called by the front end once the DLL is finished loading to finish
@@ -450,43 +450,43 @@ private:
      * value is true and the table is placed on the stack. Otherwise the function returns false and
      * nothing is put on the stack.
      */
-    bool CreateEnvironment(unsigned long api, lua_State* L, int stackLevel, int nilSentinel);
+    bool CreateEnvironment(int api, lua_State* L, int stackLevel, int nilSentinel);
 
     /**
      * Chains two tables together so that accessing members of a child table that don't exist
      * will then attempt to access them on the parent table.
      */
-    void ChainTables(unsigned long api, lua_State* L, int child, int parent);
+    void ChainTables(int api, lua_State* L, int child, int parent);
 
     /**
      *
      */
-    void CreateChainedTable(unsigned long api, lua_State* L, int nilSentinel, int localTable, int upValueTable, int globalTable);
+    void CreateChainedTable(int api, lua_State* L, int nilSentinel, int localTable, int upValueTable, int globalTable);
 
     /**
      * Sets the values of the locals at the specified stack level based on values in the table.
      * Values which are equal to the value stored at the nilSentinel stack index will be converted
      * to nils.
      */
-    void SetLocals(unsigned long api, lua_State* L, int stackLevel, int localTable, int nilSentinel);
+    void SetLocals(int api, lua_State* L, int stackLevel, int localTable, int nilSentinel);
 
     /**
      * Sets the values of the up values at the specified stack level based on values in the table.
      * Values which are equal to the value stored at the nilSentinel stack index will be converted
      * to nils.
      */
-    void SetUpValues(unsigned long api, lua_State* L, int stackLevel, int upValueTable, int nilSentinel);
+    void SetUpValues(int api, lua_State* L, int stackLevel, int upValueTable, int nilSentinel);
 
     /**
      * Sets the function (on the top of the Lua stack) to be called when a Lua
      * object is garbage collected.
      */
-    void SetGarbageCollectionCallback(unsigned long api, lua_State* L, int index);
+    void SetGarbageCollectionCallback(int api, lua_State* L, int index);
 
     /**
      * Creates a new table with weak keys or values (specified by setting the type as "k" or "v").
      */
-    void CreateWeakTable(unsigned long api, lua_State* L, const char* type);
+    void CreateWeakTable(int api, lua_State* L, const char* type);
 
     /**
      * Returns the index of the API that the VM was created in.
@@ -508,7 +508,7 @@ private:
      * Calls the function on the top of the stack when the garbage collector runs.
      * The function is popped from the stack.
      */
-    void CreateGarbageCollectionSentinel(unsigned long api, lua_State* L);
+    void CreateGarbageCollectionSentinel(int api, lua_State* L);
 
     /**
      * Used by the mechism to setup a garbage collection callback. This function is used
@@ -524,11 +524,11 @@ private:
 
     /**
      */
-    static int IndexChained(unsigned long api, lua_State* L);
+    static int IndexChained(int api, lua_State* L);
 
     /**
      */
-    static int NewIndexChained(unsigned long api, lua_State* L);
+    static int NewIndexChained(int api, lua_State* L);
 
     /**
      * Returns the end part of a file name.
@@ -538,9 +538,9 @@ private:
     /**
      * Logs information about a hook callback event. This is used for debugging.
      */
-    void LogHookEvent(unsigned long api, lua_State* L, lua_Debug* ar);
+    void LogHookEvent(int api, lua_State* L, lua_Debug* ar);
 
-    void UpdateHookMode(unsigned long api, lua_State* L, lua_Debug* hookEvent);
+    void UpdateHookMode(int api, lua_State* L, lua_Debug* hookEvent);
 
     /**
      * Calls the named meta-method for the specified value. If the value does
@@ -548,7 +548,7 @@ private:
      * Otherwise the result of calling the method is stored in the result parameter
      * (same return as lua_pcall).
      */
-    bool CallMetaMethod(unsigned long api, lua_State* L, int valueIndex, const char* method, int numResults, int& result) const;
+    bool CallMetaMethod(int api, lua_State* L, int valueIndex, const char* method, int numResults, int& result) const;
 
     /**
      * Gets the current C/C++ call stack.
@@ -559,12 +559,12 @@ private:
      * Creates a new table on the top of the stack which is the result of merging
      * the two specified tables.
      */
-    void MergeTables(unsigned long api, lua_State* L, unsigned int tableIndex1, unsigned int tableIndex2) const;
+    void MergeTables(int api, lua_State* L, unsigned int tableIndex1, unsigned int tableIndex2) const;
 
     /**
      * Gets the number of functions on the Lua stack.
      */
-    int GetStackDepth(unsigned long api, lua_State* L) const;
+    int GetStackDepth(int api, lua_State* L) const;
 
     /**
      * Returns the virtual machine that corresponds to the specified Lua state.
@@ -576,7 +576,7 @@ private:
      * Creates a call stack that unifies the native call stack and the script
      * call stack.
      */
-    unsigned int GetUnifiedStack(unsigned long api, const StackEntry nativeStack[], unsigned int nativeStackSize,
+    unsigned int GetUnifiedStack(int api, const StackEntry nativeStack[], unsigned int nativeStackSize,
         const lua_Debug scriptStack[], unsigned int scriptStackSize,
         StackEntry unifiedStack[]);
